+++ kernel/src/io.rs
@@ -1,0 +1,14 @@
+// SPDX-FileCopyrightText: 2022 Foundation Devices, Inc. <hello@foundationdevices.com>
+// SPDX-License-Identifier: Apache-2.0
+
+/// A trait for serial like drivers which are byte-oriented sinks.
+pub trait SerialWrite {
+    /// Write a single byte.
+    fn putc(&mut self, b: u8);
+}
+
+/// A trait for serial like drivers which allows reading from a source.
+pub trait SerialRead {
+    /// Read a single byte.
+    fn getc(&mut self) -> Option<u8>;
+}
